Upgrade CEF to 3.3071.1648.gb768a94 / Chromium 59.0.3071.109. This is a custom build of 3.3071.1646.gbb29707 with the OSR performance PR applied (https://bitbucket.org/chromiumembedded/cef/pull-requests/96)
This brings CasparCG up to a modern version of Chrome, and also changes the structure so that it is easier to upgrade. libcef_dll_wrapper is built locally instead of shipping a binary, as it already has an integrated CMake build system. Upgrading is now just a matter of downloading the latest version (minimal package) and renaming the unpacked directory to dependencies64/cef/windows (or same with /linux); there's no more symlinks or the likes.

Note: this commit does not compile, it just replaces the dependencies. This is to seperate the code changes into a seperate commit to be easily reviewable.